Add files via upload
This is a binlister file written in ERPLab syntax, for the EEG data published in Muller, Sirianni, & Addante, 2021 EJN, and Addante et al., 2023, Psychophysiology. It catalogs all possible combinations of events that could be interest for analyses in both memory and metacognition domains of those studies, from bin 1 through 850. For specific details on event code structure, please see Readme commit here posted on 3/26/2024