Add files via upload
Sample BinOperations files for ERPLab syntax of post-processing ERPs for individual subjects for the explicit memory context familiarity conditions reported in Addante et al., (2024/pending, Nature Communications: Psychology) paper, based upon the finding reported in Addante, Ranganath, & Yonelinas (2012, Neuroimage) paper. Note that this binoperations file was written 3/26/2024 to be specified for the conditions reported in the aforementioned 2024 manuscript. For foundational elements of the bins, see the additional files of Readme and Binlister files committed to Github 3.26.2024