Stimulus delivery files to run whole experiment
Readme File, written by Richard J. Addante, PhD, 10/15/2024, Raddante@fit.edu, Raddante@gmail.com 

Note: These studies were designed in 2016 by Richard J. Addante (RJA), coded in Presentation (Neurobehavioral Systems) in 2017 by Lindsey Sirianni, and with data collected in 2018 by Alana Muller for a Masters Thesis she completed in 2019 in Psychology at California State University – San Bernardino and published in the European Journal of Neuroscience (Muller et al., 2021). The building and design of this was supported by grants to RJA from CSUSB, and the National Institute of Health’s LRP Program via the National Institute of Neurological Disorders and Stroke. Development of code and scripts underlying the automated implicit memory analysis procedure (IMAP) for this experiment (Addante et al., 2023, Psychophysiology) was supported by grants from the NIH LRP Program of the NINDS and the Florida Institute of Technology. 

These materials were used to generate the experiment studying item and source memory, and metacognitive judgments of memory performance on a rolling basis. It has led to data that has been published thus far as:

1.	Addante, R. J., Clise, E., Waechter, R., Bengson, J., Drane, D. L., & Perez-Caban, J. (2024). A third kind of episodic memory: Context familiarity is a distinct process from item familiarity and recollection. bioRxiv/iScience, 2024.2007.2015.603640. https://www.biorxiv.org/content/10.1101/2024.07.15.603640v4.full

2.	Addante, R. J., Lopez-Calderon, J., Allen, N., Luck, C., Muller, A., Sirianni, L., Inman, C. S., & Drane, D. L. (2023). An ERP measure of non-conscious memory reveals dissociable implicit processes in human recognition using an open-source automated analytic pipeline. Psychophysiology, e14334. https://doi.org/https://doi.org/10.1111/psyp.14334  
3.	Muller, A., Sirianni, L. A., & Addante, R. J. (2021). Neural correlates of the Dunning–Kruger effect. European Journal of Neuroscience, 53(2), 460-484. https://doi.org/https://doi.org/10.1111/ejn.14935

For detailed descriptions of the paradigm’s parameters, please see the Methods sections of the manuscripts cited above. 

Description of the Experimental Materials and Files
As a brief primer, the Experiment was programmed in Presentation (Neurobehavioral Systems, Inc., https://www.neurobs.com/ ). 

We differentiated between even- and odd- numbered subjects whom received the varying conditions of order sequence of the encoding tasks (abba vs baab sequences). Thus, there are two different versions of the experiment contained in each folder, one named Even, one named Odd, and each will run that respective condition of the experiment for that kind of numbered subject ID. 
 
The details of the coding will probably need to have changes made in:
1.	the code for the directories (i.e. the paths for where the data is to be both drawn from in stimuli lists, and where the data is to be output towards), 
2.	update the event code triggers, 
3.	and update the response device and its respective codes. 
 For our studies, we use a Belkin Nostromo N50 Speedpad 10 Button Gaming Keypad, due in large part to its ergonomic design for responding on a 5-point confidence scale, and for consistency with replicating the same device used in prior studies (Addante et al., 2011, PNAS; Addante et al., 2012 Neuroimage; Addante et al., 2012 Neuropsychologia; Addante, 2015, Neuroimage; Roberts et al., 2018, Cognitive Neuroscience). 

Their buttons are ergonomic for our task, and the response keys correspond to certain codes used in Presentation, so it is a good idea to have the device for the 5-point confidence scale in memory responding.  

It is available very inexpensively used for cheap (~$15 USD): 
https://www.ebay.com/itm/266785017670?itmmeta=01HZG8W6VRQW3M3Q9YQAFEJJ8M&hash=item3e1da04f46:g:KtQAAOSwPM9mKshr&itmprp=enc%3AAQAJAAABwEW7k5tN6V4vU8s7%2F3sHy9iffRnPqTp5EsvebejlLHZ7v6d09tmrjCXsHgq6DtMVzmqJiBEjSr8oc%2Fsc7OdiSPaXjWLxKNm4%2F3huU7CySE6itYvnYY5lqsORt9s9AAPAldKsTSsldduY5M%2FNUIRHvXXiFsPRKzqdpbxC2%2B9khKECcyu%2BwVEfblK2cS1s4TVwxZRkBzeIH%2BW5Fs8pLhBob9z%2FnOOxxQ9WYIoWEf8%2BBB0XIBeXjHVGoAcJhWThoBjoJLNAUhxHA8vDIXAzBgKZ9tgYcNoX1da4Cg%2FVwEbD7KyvZTsqm4HWWsI4SwkCH8uDoSUsBaC5iQilpcXqrfWhFnaL77d0FEgOTDJ0MMS4uMZgDFGU8Y6ucfYpuEEv9V7SPyn%2FSs53pYC6xlHqZrJFKDi7JFxyetFX2SdjPKlf6gwwWITwTsiy%2FpKZnAHoOdNRyu%2F6Tka64RJ3N7NbapEpFobitpPRxp2UyZPnHMqAtOeLVvMGzl4Ba28AaioptdCtiB5s3kMAasxSwkjk71ZpXtD76psVoflYxYs%2Bbm7uc%2FvaFIRJGEuG1meTTumRcZTvsDaM5lxeBos%2BklydWIStUIs%3D%7Ctkp%3ABFBM-u3wiPxj